requirements.txt
Flask
requests
sentence-transformers
faiss-cpu
numpy
gunicorn